backend/requirements.txt
google-generativeai>=0.8.0
PyPDF2>=3.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
